Moving tests to tests.rs throughout
